# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the gateway specification for the GoCord project.",
        "title": "GoCord Gateway",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8079",
    "basePath": "/v1",
    "paths": {
...
        "/v1/guilds": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new guild",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Create a new guild",
                "parameters": [
                    {
                        "description": "Guild object",
                        "name": "guild",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/guildsv1.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/v1/guilds/{guild.id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a guild by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Get a guild by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.GetByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a guild by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Delete a guild by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Modify a guild by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Modify a guild by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Guild object",
                        "name": "guild",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/guildsv1.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/v1/guilds/{guild.id}/channels": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a guild's channels by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Get a guild's channels by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.GetGuildChannelsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new guild channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Create a new guild channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel object",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/guildsv1.CreateGuildChannelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.CreateGuildChannelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Modify a guild's channel positions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Modify a guild's channel positions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel positions object",
                        "name": "channels",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/guildsv1.ModifyGuildChannelPositionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.ModifyGuildChannelPositionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/v1/guilds/{guild.id}/members": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List a guild's members",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "List a guild's members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.ListGuildMembersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/v1/guilds/{guild.id}/members/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search a guild's members",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guilds"
                ],
                "summary": "Search a guild's members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guild.id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query string to match username(s) and nickname(s) against",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/guildsv1.SearchGuildMembersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "guildsv1.CreateGuildChannelRequest": {
            "type": "object",
            "properties": {
                "is_nsfw": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permissions": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "topic": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/guildsv1.GuildChannelType"
                },
                "user_limit": {
                    "type": "integer"
                }
            }
        },
        "guildsv1.CreateGuildChannelResponse": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/guildsv1.GuildChannel"
                }
            }
        },
        "guildsv1.CreateRequest": {
            "type": "object",
            "properties": {
                "icon": {
                    "description": "TODO limit size",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "guildsv1.CreateResponse": {
            "type": "object",
            "properties": {
                "guild": {
                    "$ref": "#/definitions/guildsv1.Guild"
                }
            }
        },
        "guildsv1.DeleteResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "guildsv1.GetByIdResponse": {
            "type": "object",
            "properties": {
                "guild": {
                    "$ref": "#/definitions/guildsv1.Guild"
                }
            }
        },
        "guildsv1.GetGuildChannelsResponse": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/guildsv1.GuildChannel"
                    }
                }
            }
        },
...
        "guildsv1.ModifyGuildChannelPositionsRequest": {
            "type": "object",
            "properties": {
                "guild_id": {
                    "type": "string"
                },
                "positions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/channelsv1.ChannelPosition"
                    }
                }
            }
        },
        "guildsv1.ModifyGuildChannelPositionsResponse": {
            "type": "object",
            "properties": {
                "positions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/channelsv1.ChannelPosition"
                    }
                }
            }
        },
...
        "guildsv1.UpdateRequest": {
            "type": "object",
            "properties": {
                "banner": {
                    "description": "TODO limit size",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "description": "TODO limit size",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "splash": {
                    "description": "TODO limit size",
                    "type": "string"
                }
            }
        },
        "guildsv1.UpdateResponse": {
            "type": "object",
            "properties": {
                "guild": {
                    "$ref": "#/definitions/guildsv1.Guild"
                }
            }
        },
        "responses.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}